secure-login-system/requirements.txt
bcrypt==4.1.2
PyJWT==2.10.1
python-dotenv==1.0.1
pyotp==2.9.0
requests==2.32.3